chore(clippy): fix lints
- remove unused cuda engine code

+++ tfhe/src/shortint/engine/server_side/bitwise_op.rs
@@ -21,7 +21,7 @@ impl ShortintEngine {
     ) -> EngineResult<()> {
         let modulus = (ct_right.degree.0 + 1) as u64;
         self.unchecked_functional_bivariate_pbs_assign(server_key, ct_left, ct_right, |x| {
-            ((x / modulus) & (x % modulus)) as u64
+            ((x / modulus) & (x % modulus))
         })?;
         ct_left.degree = ct_left.degree.after_bitand(ct_right.degree);
         Ok(())
@@ -71,7 +71,7 @@ impl ShortintEngine {
     ) -> EngineResult<()> {
         let modulus = (ct_right.degree.0 + 1) as u64;
         self.unchecked_functional_bivariate_pbs_assign(server_key, ct_left, ct_right, |x| {
-            ((x / modulus) ^ (x % modulus)) as u64
+            ((x / modulus) ^ (x % modulus))
         })?;
         ct_left.degree = ct_left.degree.after_bitxor(ct_right.degree);
         Ok(())
@@ -121,7 +121,7 @@ impl ShortintEngine {
     ) -> EngineResult<()> {
         let modulus = (ct_right.degree.0 + 1) as u64;
         self.unchecked_functional_bivariate_pbs_assign(server_key, ct_left, ct_right, |x| {
-            ((x / modulus) | (x % modulus)) as u64
+            ((x / modulus) | (x % modulus))
         })?;
         ct_left.degree = ct_left.degree.after_bitor(ct_right.degree);
         Ok(())
